+++ src/red_alert/voice_config.rs
